requirements.txt
fastapi==0.115.12
uvicorn==0.30.6
pydantic==2.11.3
openai==1.75.0
requests==2.32.3
pyyaml==6.0.2
